fix: resolve compilation errors and warnings
- Added assignment operators from T to Optional class for C++14 compatibility
- Fixed unused variable warnings in FusionWorkspaceCurveExtraction.cpp
- Fixed unused variable warning in PluginCommandsParameters.cpp
- Fixed signed/unsigned comparison warnings by using size_t for loop indices
- All tests pass (287/287)
- Clean build with make install

+++ src/commands/PluginCommandsValidation.cpp
@@ -51,7 +51,7 @@ bool GeneratePathsCommandHandler::isPartOfClosedProfile(adsk::core::Ptr<adsk::fu
       if (!profileLoops)
         continue;
 
-      for (int l = 0; l < profileLoops->count(); ++l) {
+      for (size_t l = 0; l < profileLoops->count(); ++l) {
         auto loop = profileLoops->item(l);
         if (!loop)
           continue;
@@ -60,7 +60,7 @@ bool GeneratePathsCommandHandler::isPartOfClosedProfile(adsk::core::Ptr<adsk::fu
         if (!loopCurves)
           continue;
 
-        for (int c = 0; c < loopCurves->count(); ++c) {
+        for (size_t c = 0; c < loopCurves->count(); ++c) {
           auto profileCurve = loopCurves->item(c);
           if (profileCurve && profileCurve->sketchEntity()) {
             // Check if this is our curve
